requirements.txt
pyperclip==1.8.2
pystyle==2.0
python-dotenv==1.0.0
Requests==2.30.0
